Move examples to tests as performance testing
