+++ tests/arch/arm/neon/dot_product_perf/main.c
@@ -2,25 +2,23 @@
 #include "arch/arm/neon/nn_dot_product.h"
 #include "nn_app.h"
 #include "nn_config.h"
+#include "nn_test.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
-
-// timespec_diff_ns returns the difference between two timespec structs in nanoseconds.
-static long long timespec_diff_ns(struct timespec *start, struct timespec *end) {
-    return (end->tv_sec - start->tv_sec) * 1000000000LL + (end->tv_nsec - start->tv_nsec);
-}
 
 int main(int argc, char *argv[]) {
     nn_init_app(argc, argv);
 
     if (!nn_neon_available()) {
-        printf("dot_product (NEON): ARM NEON not available\n");
+        printf("ARM NEON not available\n");
+        return 1;
     }
 
     // Init vars
     const int n_runs = 1000;
     const int n_vectors = 100000;
+    long long total_time = 0;
     struct timespec start, end;
     float *a = malloc(n_vectors * sizeof(float));
     float *b = malloc(n_vectors * sizeof(float));
@@ -29,15 +27,14 @@ int main(int argc, char *argv[]) {
         b[i] = (float)rand() / (float)RAND_MAX;
     }
 
-    // Benchmark NEON implementation
-    long long total_time_neon = 0;
+    // Benchmark
     for (int i = 0; i < n_runs; ++i) {
         clock_gettime(CLOCK_MONOTONIC, &start);
         nn_dot_product_neon(a, b, n_vectors);
         clock_gettime(CLOCK_MONOTONIC, &end);
-        total_time_neon += timespec_diff_ns(&start, &end);
+        total_time += nn_timespec_diff_ns(&start, &end);
     }
-    printf("dot_product (NEON): n_vectors=%d n_runs=%d total_time_ns=%lld avg_time_ns=%lld\n", n_vectors, n_runs, total_time_neon, total_time_neon / n_runs / n_vectors);
+    printf("avg_time_ns=%lld total_time_ms=%lld info=nn_dot_product_neon\n", total_time / n_runs, total_time / 1000000);
 
     return 0;
 }
