+++ tests/arch/generic/dot_product_perf/main.c
@@ -2,14 +2,9 @@
 #include "nn_app.h"
 #include "nn_config.h"
 #include "nn_dot_product.h"
+#include "nn_test.h"
 #include <stdio.h>
 #include <stdlib.h>
-#include <time.h>
-
-// timespec_diff_ns returns the difference between two timespec structs in nanoseconds.
-static long long timespec_diff_ns(struct timespec *start, struct timespec *end) {
-    return (end->tv_sec - start->tv_sec) * 1000000000LL + (end->tv_nsec - start->tv_nsec);
-}
 
 int main(int argc, char *argv[]) {
     nn_init_app(argc, argv);
@@ -17,6 +12,7 @@ int main(int argc, char *argv[]) {
     // Init vars
     const int n_runs = 1000;
     const int n_vectors = 100000;
+    long long total_time = 0;
     struct timespec start, end;
     float *a = malloc(n_vectors * sizeof(float));
     float *b = malloc(n_vectors * sizeof(float));
@@ -25,15 +21,14 @@ int main(int argc, char *argv[]) {
         b[i] = (float)rand() / (float)RAND_MAX;
     }
 
-    // Benchmark generic implementation
-    long long total_time_generic = 0;
+    // Benchmark
     for (int i = 0; i < n_runs; ++i) {
         clock_gettime(CLOCK_MONOTONIC, &start);
         nn_dot_product(a, b, n_vectors);
         clock_gettime(CLOCK_MONOTONIC, &end);
-        total_time_generic += timespec_diff_ns(&start, &end);
+        total_time += nn_timespec_diff_ns(&start, &end);
     }
-    printf("dot_product (generic): n_vectors=%d n_runs=%d total_time_ns=%lld avg_time_ns=%lld\n", n_vectors, n_runs, total_time_generic, total_time_generic / n_runs / n_vectors);
+    printf("avg_time_ns=%lld total_time_ms=%lld info=nn_dot_product\n", total_time / n_runs, total_time / 1000000);
 
     return 0;
 }
